Redoing Search Again
The current search just splits everything up and uses disjunctions for each term.  The only exception to that is the AND
keyword appears anywhere it splits everything up and uses the filters object instead.  Also, of higher precedence is quotes.
Quotes will lock the words next to each other.

Will definitely need automated tests to verify this!